Refactor and enhance node modification constraint checks
